refactor(db): move index file out for new globals folder
