+++ apps/bot/src/types/command/Command.ts
@@ -1,6 +1,6 @@
 import type { ColorResolvable, Message } from "discord.js"
 import type { Payload } from "payload"
-import type { db } from "@/database/collections"
+import type { db } from "@/database"
 import type { Client } from "../Client"
 import type {
   DmsOnlyPermissionSet,
